refactor(views): Component, Observer import 경로 수정
