+++ src/views/components/core/Component.js
