+++ src/views/ground/board/content/content.js
@@ -1,9 +1,9 @@
-import Component from '../../../components/common/Component.js';
+import Component from '../../../components/core/Component.js';
 import Content from '../../../components/content/content.js';
+import { observable, observe } from '../../../components/core/observer.js';
 import * as api from '../../../public/api.js';
 import ErrorModal from '../../../components/common/ErrorModal.js';
 import ErrorPage from '../../../components/common/errorPage.js';
-import { observable, observe } from '../../../components/common/observer.js';
 import { icons } from '../../../public/icons.js';
 import { isNull } from '../../../public/util.js';
