+++ src/views/ground/update/content/content.js
@@ -1,8 +1,8 @@
-import Component from '../../../components/common/Component.js';
+import Component from '../../../components/core/Component.js';
 import Content from '../../../components/content/content.js';
+import { observable, observe } from '../../../components/core/observer.js';
 import * as api from '../../../public/api.js';
 import { RegEx, isEmpty, isNull } from '../../../public/util.js';
-import { observable, observe } from '../../../components/common/observer.js';
 import ErrorModal from '../../../components/common/ErrorModal.js';
 import SuccessModal from '../../../components/common/SuccessModal.js';
 import { icons } from '../../../public/icons.js';
